[proxy][functions] Issue #2154: proxy should be able to forward rest requests to function workers cluster
*Motivation*

Function workers can be deployed as a separate cluster. If so, proxy is not able to forward the functions
related rest calls to the correct server.

*Changes*

Add two settings in proxy configuration to allow proxy configuring forwarding functions related rest calls
to function worker cluster.

*Tests*

Verified with changes in integration tests (manually). It is hard to add the integration tests based on
current integration tests. will add them in a separate PR.